+++ src/ArrayComparator.php
@@ -24,7 +24,6 @@ use function sprintf;
 use function str_replace;
 use function trim;
 use function usort;
-use SebastianBergmann\Exporter\Exporter;
 
 /**
  * @no-named-arguments Parameter names are not covered by the backward compatibility promise for sebastian/comparator
@@ -70,7 +69,7 @@ class ArrayComparator extends Comparator
         $actualAsString   = "Array (\n";
         $expectedAsString = "Array (\n";
         $equal            = true;
-        $exporter         = new Exporter;
+        $exporter         = $this->exporter();
 
         foreach ($expected as $key => $value) {
             unset($remaining[$key]);