+++ src/Factory.php
@@ -13,6 +13,7 @@ use const PHP_VERSION;
 use function array_unshift;
 use function extension_loaded;
 use function version_compare;
+use SebastianBergmann\Exporter\Exporter;
 
 /**
  * @no-named-arguments Parameter names are not covered by the backward compatibility promise for sebastian/comparator
@@ -34,6 +35,7 @@ final class Factory
     /** @var positive-int */
     private int $contextLines               = 3;
     private bool $closureComparisonOccurred = false;
+    private Exporter $exporter;
 
     public static function getInstance(): self
     {
@@ -46,6 +48,8 @@ final class Factory
 
     public function __construct()
     {
+        $this->exporter = new Exporter;
+
         $this->registerDefaultComparators();
     }
 
@@ -63,6 +67,16 @@ final class Factory
     public function setContextLines(int $contextLines): void
     {
         $this->contextLines = $contextLines;
+    }
+
+    public function exporter(): Exporter
+    {
+        return $this->exporter;
+    }
+
+    public function setExporter(Exporter $exporter): void
+    {
+        $this->exporter = $exporter;
     }
 
     /**